handling invalid settings
