+++ lib/src/generators/assets_generator.dart
@@ -8,14 +8,17 @@ import '../settings/asset_type.dart';
 import '../settings/flutter.dart';
 import '../settings/flutter_gen.dart';
 import '../utils/camel_case.dart';
+import '../utils/error.dart';
 import 'generator_helper.dart';
 import 'integrations/integration.dart';
 import 'integrations/svg_integration.dart';
 
 String generateAssets(File pubspecFile, DartFormatter formatter,
     FlutterGen flutterGen, FlutterAssets assets) {
-  assert(assets != null && assets.hasAssets,
-      throw 'The value of "flutter/assets:" is incorrect.');
+  if (assets == null || !assets.hasAssets) {
+    throw InvalidSettingsException(
+        'The value of "flutter/assets:" is incorrect.');
+  }
 
   final importsBuffer = StringBuffer();
   final classesBuffer = StringBuffer();
