+++ lib/src/generators/fonts_generator.dart
@@ -5,11 +5,14 @@ import 'package:yaml/yaml.dart';
 import '../settings/flutter.dart';
 import '../utils/camel_case.dart';
 import '../utils/cast.dart';
+import '../utils/error.dart';
 import 'generator_helper.dart';
 
 String generateFonts(DartFormatter formatter, FlutterFonts fonts) {
-  assert(fonts != null && fonts.hasFonts,
-      throw 'The value of "flutter/fonts:" is incorrect.');
+  if (fonts == null || !fonts.hasFonts) {
+    throw InvalidSettingsException(
+        'The value of "flutter/fonts:" is incorrect.');
+  }
 
   final buffer = StringBuffer();
   buffer.writeln(header);
