+++ lib/src/utils/error.dart
@@ -1,0 +1,5 @@
+class InvalidSettingsException implements Exception {
+  const InvalidSettingsException(this.message);
+
+  final String message;
+}
